refactor(foundation): move MockEnvironmentName + PromotionEntityType (A10)
Phase 6 / Sub-phase A10: Extract the simple workspace promotion enum types to
foundation. These are pure-data 3-variant enums with display/parsing helpers.

Foundation:
- workspace_promotion::MockEnvironmentName (Dev, Test, Prod) + helpers
- workspace_promotion::PromotionEntityType (Scenario, Persona, Config) + Display
- async-trait dep added
- uuid feature "serde,v4" enabled

Core re-export shims:
- workspace/mock_environment.rs: MockEnvironmentName re-exported
- workspace/scenario_promotion.rs: PromotionEntityType re-exported
- (PromotionRequest, PromotionStatus, PromotionService trait stay in core
  because their richer field shapes don't translate cleanly to foundation)

Migrated mockforge-pipelines/src/steps/auto_promote.rs:
- PromotionEntityType from foundation
- MockEnvironmentName::from_str() from foundation
- Added mockforge-foundation Cargo dep

Workspace builds clean.
