+++ crates/mockforge-foundation/src/lib.rs
@@ -16,6 +16,7 @@ pub mod encryption_error;
 pub mod error;
 pub mod intelligent_behavior;
 pub mod state_machine;
+pub mod workspace_promotion;
 
 pub use encryption_error::{EncryptionError, EncryptionResult};
 pub use error::{Error, Result};
